Lowercase filename + static class to functions
Static classes turned into a module of functions, filenames made lowercase, initialize function moved into __init__.py and BaseServer now callable as application
