+++ setup.cfg
@@ -13,10 +13,9 @@ packages =
     sserver
     sserver.crypto
     sserver.endpoint
-    sserver.file
-    sserver.log
     sserver.mixin
-    sserver.tool
+    sserver.path
+    sserver.util
 package_dir =
     =src
 include_package_data = True
