+++ src/sserver/__init__.py
@@ -1,0 +1,5 @@
+from sserver.util import config
+from sserver.util import route
+
+config.load()
+route.load()
