--- src/sserver/log/__init__.py
@@ -1,1 +1,0 @@
-"""Provides modules to handle logging."""
