+++ src/sserver/util/exception.py
